fix mouse input truncation with non-integral sensitivity (ironwail)

+++ Quake/in_sdl.c
@@ -864,7 +864,7 @@ void IN_JoyMove (usercmd_t *cmd)
 
 void IN_MouseMove(usercmd_t *cmd)
 {
-	int		dmx, dmy;
+	float		dmx, dmy; // woods fix mouse input truncation with non-integral sensitivity (ironwail v.0.5.0)
 
 	dmx = total_dx * sensitivity.value;
 	dmy = total_dy * sensitivity.value;
